// src/app/components/layout/personal-cabinet/card-form/card-form.interface.ts
export interface IErrorsCard {
  month: number
  year: number
  first_input: string
  second_input: string
  third_input: string
  fourth_input: string
  cvv: string
}

export interface IRegisterCard {
  month: number
  first_input: string
  second_input: string
  third_input: string
  fourth_input: string
  year: number
  cvv: string
}